requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
azure-functions
azure-identity
azure-mgmt-storage
azure-mgmt-resource
azure-mgmt-monitor
azure-mgmt-compute 
azure-mgmt-containerregistry
azure-mgmt-containerinstance
azure-storage-blob
azure-messaging-webpubsubclient
azure-messaging-webpubsubservice
azure-mgmt-appcontainers